fix: extract components and fix bugs
