+++ apps/builder/src/middleware/reduxAsync/sendMessage/index.ts
@@ -1,8 +1,8 @@
 import { PayloadAction } from "@reduxjs/toolkit"
-import { appInfoAsync } from "@/middleware/reduxAsync/sendMessage/appInfoMethod"
 import { getCurrentTeamInfo } from "@/redux/team/teamSelector"
 import { RootState } from "@/store"
 import { actionsAsync } from "./actionMethod"
+import { appInfoAsync } from "./appInfoMethod"
 import { appsAsync } from "./appsMethod"
 import { componentsAsync } from "./componentsMethod"
 import { dottedLineSquareAsync } from "./dottedLineSquareMethod"
