+++ apps/builder/src/page/App/components/PageNavBar/AppName/style.ts
@@ -1,0 +1,12 @@
+import { css } from "@emotion/react"
+import { getColor } from "@illa-design/react"
+
+export const triggerStyle = css`
+  margin-top: 20px;
+`
+
+export const nameStyle = css`
+  font-weight: 500;
+  cursor: pointer;
+  color: ${getColor("grayBlue", "02")};
+`
