+++ apps/builder/src/page/App/components/PageNavBar/interface.ts
@@ -2,9 +2,12 @@ import { HTMLAttributes } from "react"
 
 export interface PageNavBarProps extends HTMLAttributes<HTMLDivElement> {}
 
-export interface PreviewPopContentProps {
-  viewportWidth?: number
-  viewportHeight?: number
+export type AppSizeType = "fluid" | "desktop" | "tablet" | "custom"
+
+export interface AppNameEditorModalProps {
+  onSuccess: () => void
 }
 
-export type AppSizeType = "fluid" | "desktop" | "tablet" | "custom"
+export interface AppNameProps {
+  appName: string
+}
